amplify-infra/requirements.txt
attrs==20.3.0
aws-cdk.assets==1.105.0
aws-cdk.aws-amplify==1.105.0
aws-cdk.aws-applicationautoscaling==1.105.0
aws-cdk.aws-autoscaling-common==1.105.0
aws-cdk.aws-cloudwatch==1.105.0
aws-cdk.aws-codebuild==1.105.0
aws-cdk.aws-codecommit==1.105.0
aws-cdk.aws-codeguruprofiler==1.105.0
aws-cdk.aws-ec2==1.105.0
aws-cdk.aws-ecr==1.105.0
aws-cdk.aws-ecr-assets==1.105.0
aws-cdk.aws-efs==1.105.0
aws-cdk.aws-events==1.105.0
aws-cdk.aws-iam==1.105.0
aws-cdk.aws-kms==1.105.0
aws-cdk.aws-lambda==1.105.0
aws-cdk.aws-logs==1.105.0
aws-cdk.aws-s3==1.105.0
aws-cdk.aws-s3-assets==1.105.0
aws-cdk.aws-sam==1.105.0
aws-cdk.aws-secretsmanager==1.105.0
aws-cdk.aws-signer==1.105.0
aws-cdk.aws-sqs==1.105.0
aws-cdk.aws-ssm==1.105.0
aws-cdk.cloud-assembly-schema==1.105.0
aws-cdk.core==1.105.0
aws-cdk.cx-api==1.105.0
aws-cdk.region-info==1.105.0
cattrs==1.5.0
constructs==3.3.75
jsii==1.29.0
publication==0.0.3
python-dateutil==2.8.1
six==1.16.0
typing-extensions==3.10.0.0
